refactor(whitelist-preconfirmation-driver): centralize tx-list decode logic and dedupe ingress envelope ingest
