+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/api/service/tests.rs
@@ -3,14 +3,12 @@ use std::io::Write;
 use alloy_primitives::Address;
 use flate2::{Compression, write::ZlibEncoder};
 
-use super::{
-    compression::decompress_tx_list,
-    lookahead::{is_fee_recipient_allowed_for_slot, slot_matches_range},
-};
+use super::lookahead::{is_fee_recipient_allowed_for_slot, slot_matches_range};
 use crate::{
     api::types::{LookaheadStatus, SlotRange},
     error::WhitelistPreconfirmationDriverError,
     importer::{MAX_COMPRESSED_TX_LIST_BYTES, MAX_DECOMPRESSED_TX_LIST_BYTES},
+    tx_list::decompress_tx_list,
 };
 
 fn compress(payload: &[u8]) -> Vec<u8> {
