+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/importer/mod.rs
@@ -33,14 +33,11 @@ mod validation;
 #[cfg(test)]
 mod tests;
 
-pub(crate) use validation::validate_execution_payload_for_preconf;
-
 /// Maximum compressed tx-list size accepted from a preconfirmation payload.
-pub(crate) const MAX_COMPRESSED_TX_LIST_BYTES: usize = 131_072 * 6;
+pub(crate) use crate::tx_list::MAX_COMPRESSED_TX_LIST_BYTES;
 /// Maximum decompressed tx-list size accepted from a preconfirmation payload.
-///
-/// Align with the preconfirmation tx-list cap to avoid zlib bomb expansion on untrusted payloads.
-pub(crate) const MAX_DECOMPRESSED_TX_LIST_BYTES: usize = 8 * 1024 * 1024;
+pub(crate) use crate::tx_list::MAX_DECOMPRESSED_TX_LIST_BYTES;
+pub(crate) use validation::validate_execution_payload_for_preconf;
 /// Dependency bundle for constructing [`WhitelistPreconfirmationImporter`].
 pub(crate) struct WhitelistPreconfirmationImporterParams<P>
 where
