+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/lib.rs
@@ -11,6 +11,7 @@ pub mod metrics;
 mod network;
 mod preconf_ingress_sync;
 mod runner;
+mod tx_list;
 mod whitelist_fetcher;
 
 pub use error::{Result, WhitelistPreconfirmationDriverError};
